requirements-dev.txt
-r requirements.txt
pytest>=8.0.0
pytest-cov>=5.0.0
moto[ec2,cloudwatch,pricing]>=5.0.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.10.0
boto3-stubs[ec2,cloudwatch,pricing]>=1.34.0
